fix: Rename azure plugin to azuredevops
The name conflicts with existing go plugin
