+++ backend/python/plugins/azuredevops/azuredevops/streams/builds.py
@@ -17,9 +17,9 @@ from typing import Iterable
 
 import iso8601 as iso8601
 
-from azure.api import AzureDevOpsAPI
-from azure.models import AzureDevOpsConnection, GitRepository
-from azure.models import Build
+from azuredevops.api import AzureDevOpsAPI
+from azuredevops.models import AzureDevOpsConnection, GitRepository
+from azuredevops.models import Build
 from pydevlake import Context, DomainType, Stream, logger
 import pydevlake.domain_layer.devops as devops
 
@@ -31,8 +31,8 @@ class Builds(Stream):
     def collect(self, state, context) -> Iterable[tuple[object, dict]]:
         connection: AzureDevOpsConnection = context.connection
         repo: GitRepository = context.scope
-        azure_api = AzureDevOpsAPI(connection.base_url, connection.pat)
-        response = azure_api.builds(repo.org_id, repo.project_id, repo.id, repo.provider)
+        azuredevops_api = AzureDevOpsAPI(connection.base_url, connection.pat)
+        response = azuredevops_api.builds(repo.org_id, repo.project_id, repo.id, repo.provider)
         for raw_build in response:
             yield raw_build, state
